+++ src/command/completion/script/embed.go
@@ -7,4 +7,6 @@ import (
 var (
 	//go:embed bash.sh
 	Bash string
+	//go:embed zsh.sh
+	Zsh string
 )